% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robregcc.R
\name{classo}
\alias{classo}
\title{Estimate parameters of linear regression model with compositional covariates using method suggested by Pixu shi.}
\usage{
classo(Xt, y, C, we = NULL, type = 1, control = list())
}
\arguments{
\item{Xt}{CLR transformed predictor matrix.}

\item{y}{model response vector}

\item{C}{sub-compositional matrix}

\item{we}{specify weight of model parameter}

\item{type}{1/2 for l1 / l2 loss in the model}

\item{control}{a list of internal parameters controlling the model fitting}
}
\value{
\item{beta}{model parameter estimate}
}
\description{
The model uses scaled lasoo approach for model selection.
}
\examples{
\dontrun{
### specify examples here to be shown in the package:
print("Examples")
}
}
\references{
Shi, P., Zhang, A. and Li, H., 2016. \emph{Regression analysis for microbiome compositional data. The Annals of Applied Statistics}, 10(2), pp.1019-1040.
}
