% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{coef_cc}
\alias{coef_cc}
\alias{coef_cc.robregcc}
\title{Extract coefficients estimate from the sparse version of the robregcc fitted object.}
\usage{
coef_cc(object, ...)

coef_cc.robregcc(object, type = 0, s = 1)
}
\arguments{
\item{object}{Object generated by \code{robregcc}.}

\item{...}{additional argument}

\item{type}{0/1 residual estimate before/after sanity check}

\item{s}{0/1 no/yes 1se rule}
}
\value{
coefficient estimate
}
\description{
S3 methods extracting estimated coefficients for objects generated by
\code{robregcc}. Robust coeffcient estimate.
}
\examples{
 
\dontrun{
# fit.ada, fit.soft, fit.hard are fitted object from the sparse version of the robregcc model
coef_cc(fit.ada)
coef_cc(fit.soft)
coef_cc(fit.hard)
}
}
