% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robregcc.R
\name{cpsc_sp}
\alias{cpsc_sp}
\title{Principal sensitive component analysis with compositional covariates in sparse setting.}
\usage{
cpsc_sp(X0, y0, alp = 0.4, cfac = 2, b1 = 0.25, cc1 = 2.937,
  C = NULL, we, type, control = list())
}
\arguments{
\item{X0}{CLR transformed predictor matrix.}

\item{y0}{model response vector}

\item{alp}{(0,0.5) fraction of data sample to be removed to generate subsample}

\item{cfac}{initial value of shift parameter for weight construction/initialization}

\item{b1}{tukey bisquare function parameter producing desired breakdown point}

\item{cc1}{tukey bisquare function parameter producing desired breakdown point}

\item{C}{sub-compositional matrix}

\item{we}{penalization index for model parameters beta}

\item{type}{1/2 for l1 / l2 loss in the model}

\item{control}{a list of internal parameters controlling the model fitting}
}
\value{
\item{betaf}{TModel parameter estimate}
  \item{residuals}{residual estimate}
}
\description{
Produce model and its residual estimate based on PCS analysis.
}
\examples{
 
\dontrun{
### specify examples here to be shown in the package:
print("Examples")
}
}
\references{
Mishra, A., Mueller, C.,(2018) \emph{Robust regression with compositional covariates. In prepration}.
}
