% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getscsfun}
\alias{getscsfun}
\title{Subfunction for principal sensitive component analysis:}
\usage{
getscsfun(Xa, ya, alp0 = 0.4, b1 = 0.25, cc1 = 2.937)
}
\arguments{
\item{Xa}{CLR transformed predictor matrix.}

\item{ya}{model response vector}

\item{alp0}{(0,0.5) fraction of data sample to be removed to generate subsample}

\item{b1}{tukey bisquare function parameter producing desired breakdown point}

\item{cc1}{tukey bisquare function parameter producing desired breakdown point}
}
\value{
\item{beta}{Model parameter estimate}
  \item{scale}{scale estimate}
}
\description{
Subfubction PCS non-sparse.
}
\examples{
### specify examples here to be shown in the package:
print("aditya")
}
\references{
Mishra, A., Mueller, C.,(2018) \emph{Robust regression with compositional covariates. In prepration}.
}
