% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getscsfun.sp}
\alias{getscsfun.sp}
\title{Subfunction for principal sensitive component analysis (sparsity):}
\usage{
getscsfun.sp(Xa, ya, alp0 = 0.4, b1 = 0.25, cc1 = 2.937, C = NULL,
  we, type, control = list())
}
\arguments{
\item{Xa}{CLR transformed predictor matrix.}

\item{ya}{model response vector}

\item{alp0}{(0,0.5) fraction of data sample to be removed to generate subsample}

\item{b1}{tukey bisquare function parameter producing desired breakdown point}

\item{cc1}{tukey bisquare function parameter producing desired breakdown point}

\item{C}{sub-compositional matrix}

\item{we}{penalization index for model parameters beta}

\item{type}{1/2 for l1 / l2 loss in the model}

\item{control}{a list of internal parameters controlling the model fitting}
}
\value{
\item{beta}{Model parameter estimate}
  \item{scale}{scale estimate}
}
\description{
Subfubction PCS sparse.
}
\examples{
### specify examples here to be shown in the package:
print("aditya")
}
\references{
Mishra, A., Mueller, C.,(2018) \emph{Robust regression with compositional covariates. In prepration}.
}
