% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_robregcc.R
\name{plot_cv}
\alias{plot_cv}
\alias{plot_cv.robregcc}
\title{Plot cross-validation error plot}
\usage{
plot_cv(object)

plot_cv.robregcc(object)
}
\arguments{
\item{object}{robregcc fitted onject}
}
\value{
generate cv error plot
}
\description{
S3 methods plotting crossvalidation error using the object obtained from \code{robregcc}.
}
\examples{
 
 
\dontrun{
# fit.ada, fit.soft, fit.hard are fitted object from the sparse version of the robregcc model

plot_cv(fit.ada)
plot_cv(fit.soft)
plot_cv(fit.hard)
}
}
