% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_robregcc.R
\name{plot_path}
\alias{plot_path}
\alias{plot_path.robregcc}
\title{Plot solution path at different value of lambda}
\usage{
plot_path(object, ptype = 0)

plot_path.robregcc(object, ptype = 0)
}
\arguments{
\item{object}{Object generated by \code{robregcc}.}

\item{ptype}{path type 0/1 for Gamma/Beta path respectvely}
}
\value{
plot solution path
}
\description{
S3 methods plotting solution path of model parameter and mean shift using the object obtained from \code{robregcc}.
}
\examples{
 
\dontrun{
# fit.ada, fit.soft, fit.hard are fitted object from the sparse version of the robregcc model

plot_path(fit.ada)
plot_path(fit.soft)
plot_path(fit.hard)
}
}
