% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_robregcc.R
\name{plot_resid}
\alias{plot_resid}
\alias{plot_resid.robregcc}
\title{Plot residuals estimate from robregcc object}
\usage{
plot_resid(object, type = 0, s = 1)

plot_resid.robregcc(object, type = 1, s = 1)
}
\arguments{
\item{object}{Object generated by \code{robregcc}.}

\item{type}{0/1 residual estimate before/after sanity check}

\item{s}{0/1 no/yes 1se rule}
}
\value{
plot estimated residual
}
\description{
S3 methods extracting residuals from the objects generated by
\code{robregcc}.
}
\examples{
 
\dontrun{
### specify examples here to be shown in the package:
print("Examples")
}
}
