% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{residuals}
\alias{residuals}
\alias{residuals.robregcc}
\title{Extract residuals estimate from the sparse version of the robregcc fitted object.}
\usage{
\method{residuals}{robregcc}(object, ...)
}
\arguments{
\item{object}{robregcc fitted onject}

\item{...}{Other argumnts for future usage.}
}
\value{
residuals estimate
}
\description{
Robust residuals estimate
}
\examples{
 
\dontrun{
# fit.ada, fit.soft, fit.hard are fitted object from the sparse version of the robregcc model
residuals(fit.ada)
residuals(fit.soft)
residuals(fit.hard)
}
}
