% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robregcc.R
\name{robregcc_nsp}
\alias{robregcc_nsp}
\title{Robust model estimation approach for regression with compositional covariates.}
\usage{
robregcc_nsp(X, y, C, intercept = FALSE, gamma.wt = NULL,
  control = list(), penalty.index = 1)
}
\arguments{
\item{X}{CLR transformed predictor matrix.}

\item{y}{model response vector}

\item{C}{sub-compositional constraint matrix}

\item{intercept}{true/false to include intercept term in the model}

\item{gamma.wt}{initial value of shift parameter for weight construction/initialization}

\item{control}{a list of internal parameters controlling the model fitting}

\item{penalty.index}{1, 2, 3 corresponding to adaptive, soft and hard penalty}
}
\value{
\item{Method}{Type of penalty used}
  \item{betapath}{model parameter estimate along solution path}
  \item{gammapath}{shift parameter estimate along solution path}
  \item{lampath}{sequence of fitted lambda)}
  \item{X}{predictors}
  \item{y}{response}
}
\description{
Generate solution path for range of lambda in case of where model parameter beta is not assumed to be sparse(nsp).
}
\examples{
\dontrun{
### specify examples here to be shown in the package:
print("Examples")
}
}
\references{
Mishra, A., Mueller, C.,(2018) \emph{Robust regression with compositional covariates. In prepration}.
}
