% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robregcc.R
\name{robregcc_sim}
\alias{robregcc_sim}
\title{Simulation data}
\usage{
robregcc_sim(n, betacc, O, Sigma, levg, snr, shft, m, C, out = list())
}
\arguments{
\item{n}{sample size}

\item{betacc}{model parameter satisfying compositional covariates}

\item{O}{number of outlier}

\item{Sigma}{covariance matrix of simulated predictors}

\item{levg}{1/0 whether to include leveraged observation or not}

\item{snr}{noise to signal ratio}

\item{shft}{multiplying factor to model variance for creating outlier}

\item{m}{test sample size}

\item{C}{subcompositional matrix}

\item{out}{list for obtaining output with simulated data structure}
}
\value{
a list containing simulated output.
}
\description{
Simulate data for the robust regression with compositional covariates
}
\examples{
 
\dontrun{
### specify examples here to be shown in the package:
print("Examples")
}
}
