% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robregcc.R
\name{robregcc_sp}
\alias{robregcc_sp}
\title{Robust model estimation approach for regression with compositional covariates.}
\usage{
robregcc_sp(X, y, C, beta.init = NULL, gamma.init = NULL,
  beta.wt = NULL, gamma.wt = NULL, control = list(),
  penalty.index = 3, alpha = 1)
}
\arguments{
\item{X}{CLR transformed predictor matrix. Centered X}

\item{y}{model response vector, centered y}

\item{C}{sub-compositional matrix}

\item{beta.init}{initial value of model parameter beta}

\item{gamma.init}{inital value of shift parameter gamma}

\item{beta.wt}{specify weight of model parameter}

\item{gamma.wt}{initial value of shift parameter for weight construction/initialization}

\item{control}{a list of internal parameters controlling the model fitting}

\item{penalty.index}{a vector of length 2 specifying type of penalty for model parameter and shift parameter respectively. 1, 2, 3 corresponding to adaptive, soft and hard penalty}

\item{alpha}{elastic net penalty}
}
\value{
\item{Method}{Type of penalty used}
  \item{betapath}{model parameter estimate along solution path}
  \item{gammapath}{shift parameter estimate along solution path}
  \item{lampath}{sequence of fitted lambda)}
  \item{k0}{scaling factor}
}
\description{
Fit regression model with compositional covariates for a  range of tuning parameter lambda. Model parameters is assumed to be sparse.
}
\examples{
\dontrun{
### specify examples here to be shown in the package:
print("Examples")
}
}
\references{
Mishra, A., Mueller, C.,(2018) \emph{Robust regression with compositional covariates. In prepration}.
}
