% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulate_robregcc}
\alias{simulate_robregcc}
\title{Simulated date for testing functions in the robregcc package (sparse setting).}
\format{
A list with three components: 
\describe{
  \item{X}{Compositional predictors.}
  \item{y}{Outcome with outliers.}
  \item{C}{Sub-cmposition matrix.}
}
}
\source{
Similated data
}
\usage{
data(simulate_robregcc)
}
\description{
A list of response (y), predictors (X) and sub-cpmposition matrix (C).
}
\details{
Vector y, response with a certain percentage of observations as outliers.

Matrix X, Compositional predictors.
}
\examples{

library(robregcc)
data(simulate_robregcc)
X <- simulate_robregcc$X;
y <- simulate_robregcc$y
C <- simulate_robregcc$C
n <- nrow(X); p <- ncol(X); k <-  nrow(C)

}
\keyword{datasets}
