% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\docType{data}
\name{simulate_robregcc_sp}
\alias{simulate_robregcc_sp}
\title{Simulated date for testing functions in the robregcc package (sparse setting).}
\format{A list with three components: 
\describe{
  \item{X}{Compositional predictors.}
  \item{y}{Outcome with outliers.}
  \item{C}{Sub-cmposition matrix.}
}}
\usage{
data(simulate_robregcc_sp)
}
\description{
A list of response (y), predictors (X) and sub-cpmposition matrix (C).
}
\details{
Vector y, response with a certain percentage of observations as outliers.

Matrix X, Compositional predictors.
}
\examples{
data(simulate_robregcc_sp)
X <- simulate_robregcc_sp$X;
y <- simulate_robregcc_sp$y
C <- simulate_robregcc_sp$C
n <- nrow(X); p <- ncol(X); k <-  nrow(C)
}
\keyword{datasets}
