% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{coef_cc}
\alias{coef_cc}
\alias{coef_cc.robregcc}
\title{Estimated coefficients}
\usage{
coef_cc(object, ...)

coef_cc.robregcc(object, type = 0, s = 1)
}
\arguments{
\item{object}{Object generated by \code{robregcc}.}

\item{...}{additional argument}

\item{type}{0/1 residual estimate before/after sanity check}

\item{s}{0/1 no/yes 1se rule}
}
\value{
coefficient estimate
}
\description{
S3 methods extracting estimated coefficients for objects generated by
\code{robregcc}. Robust coeffcient estimate.
}
